Refactor: Move task controller to API namespace and add parent_id support
