+++ backend/config/routes.rb
@@ -1,5 +1,7 @@
 Rails.application.routes.draw do
+  namespace :api do
+    resources :tasks
+  end
+
   devise_for :users
-  resources :tasks
-  root "tasks#index"
 end
